Primera versión para pro

+++ src/Service/PostHandlerService.php
@@ -4,23 +4,29 @@
 namespace App\Service;
 
 use App\Entity\Post;
+use App\Entity\PostType;
 use App\Form\Model\PostDto;
+use App\Repository\PostRepository;
 use App\Repository\PostTypeRepository;
 use App\Repository\ShopRepository;
 use App\Utils\Constants;
 use Doctrine\DBAL\Exception as DoctrineException;
 use Doctrine\ORM\EntityManagerInterface;
 use Exception;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\UrlHelper;
 
 class PostHandlerService
 
 {
-      private EntityManagerInterface $entityManager;
-      private PostTypeRepository $typeRepository;
-      private ShopRepository  $shopRepository;
-      private ImageHandlerService $imageService;
-      private UrlHelper $urlHelper;
+    private EntityManagerInterface $entityManager;
+    private PostTypeRepository $typeRepository;
+    private ShopRepository  $shopRepository;
+    private ImageHandlerService $imageService;
+    private PostRepository  $postRepository;
+    private UrlHelper $urlHelper;
+
+
 
     /**
      * PostHandlerService constructor.
@@ -30,15 +36,46 @@ class PostHandlerService
      * @param ImageHandlerService $imageService
      * @param UrlHelper $urlhelper
      */
-    public function __construct(EntityManagerInterface $entityManager, PostTypeRepository $typeRepository, ShopRepository $shopRepository, ImageHandlerService $imageService, UrlHelper $urlhelper)
+    public function __construct(EntityManagerInterface $entityManager,PostRepository  $postRepository, PostTypeRepository $typeRepository, ShopRepository $shopRepository, ImageHandlerService $imageService, UrlHelper $urlhelper)
     {
         $this->entityManager = $entityManager;
         $this->typeRepository = $typeRepository;
         $this->shopRepository = $shopRepository;
         $this->imageService = $imageService;
         $this->urlHelper = $urlhelper;
+        $this->postRepository =  $postRepository;
     }
 
+
+    public function createPostFromRequest(PostDto $postDto): Response
+    {
+
+        $typeFromDb = $this->typeRepository->findOneBy(['type' => $postDto->getTypeValue()] ) ;
+
+        $type = new PostType();
+
+        if (!$typeFromDb) {
+            $type->setType($postDto->getTypeValue());
+            $this->entityManager->persist($type);
+            $this->entityManager->flush();
+            $typeFromDb = $type;
+        }
+
+        $shop = $this->shopRepository->find($this->shopRepository->findOneBy(['uid' => $postDto->getShopUid()] ));
+
+        if(!$shop) {
+            return new Response('ERROR SHOP NOT FOUND.',  Response::HTTP_NOT_FOUND);
+        }
+
+        $fileNameImage = $this->imageService->saveImage($postDto->getImage(),Constants::postImageDirectory );
+        $postDto->setImage($this->urlHelper->getAbsoluteUrl(constants::pathOfImagesByDefault. $fileNameImage));
+
+        if($this->persistPost(PostDto::createEntityFromRequest($postDto , $typeFromDb, $shop))){
+            return  new Response('Post created .',  Response::HTTP_CREATED);
+        }
+
+        return new Response('ERROR AL CREAR EL POST',  Response::HTTP_BAD_REQUEST);
+    }
 
     public function persistPost(Post $post): bool
     {
@@ -48,32 +85,29 @@ class PostHandlerService
             return true;
         }catch (Exception  $ex){
             return false;
-    }
-
-    }
-
-    public function createPostFromRequest(PostDto $postDto): Post
-    {
-        $type = $this->typeRepository->find($postDto->getTypeId());
-        $shop = $this->shopRepository->find($postDto->getShopRelatedId());
-        $fileNameImage = $this->imageService->saveImage($postDto->getImage(),Constants::postImageDirectory );
-        $postDto->setImage($this->urlHelper->getAbsoluteUrl(constants::pathOfImagesByDefault. $fileNameImage));
-        if(!$type || !$shop) throw new Exception('DATA NOT FOUND');
-
-        return PostDto::createEntityFromRequest($postDto , $type , $shop);
-    }
-
-    public function deletePost(Post $post): Post
-    {
-        try{
-            $this->imageService->deleteImage($post->getImage());
-            $this->entityManager->remove($post);
-            $this->entityManager->flush();
-        }catch (Exception $exception){
-            throw new DoctrineException("Error try to remove Post");
         }
-        return $post;
     }
+
+    public function deletePost(int $postId): Response
+    {
+
+        try{
+            $post =  $this->postRepository->find($postId);
+
+            if($post){
+                $this->imageService->deleteImage($post->getImage());
+                $this->entityManager->remove($post);
+                $this->entityManager->flush();
+                return new Response('POST ELIMINADO',  Response::HTTP_OK);
+            }
+            return  new Response('POST NO ENCONTRADO',  Response::HTTP_NOT_FOUND);
+
+        }catch (Exception $exception){
+            return new Response('ERROR ELIMINAR EL POST',  Response::HTTP_NOT_MODIFIED);
+        }
+    }
+
+
 
 
 }
